Base64 encoded values in plain text detection (#478)
* Base64 search for LS0t 2023-12-17T11:43:50+02:00

* custom benchmark for test

* style

* unused var optimized

* BM scores fix

* Refactoring logic for using filters for base64pem

* style

* Update .github/workflows/benchmark.yml
