+++ cicd/benchmark.txt
@@ -1,4 +1,4 @@
-DATA: 19419621 valid lines. MARKUP: 74230 items
+DATA: 19419621 valid lines. MARKUP: 74231 items
 Category                      Positives    Negatives    Template
 --------------------------  -----------  -----------  ----------
 Authentication Key & Token           77            1          31
@@ -7,19 +7,19 @@ Generic Token                       333           45         558
 Other                               844        63294         635
 Password                           1402          110        4164
 Predefined Pattern                  330            2          40
-Private Key                        1008            2          32
+Private Key                        1009            2          32
 Seed, Salt, Nonce                    40            4           4
-TOTAL:                             5090        63473        5667
-Detected Credentials: 5864
-credsweeper result_cnt : 5161, lost_cnt : 0, true_cnt : 4263, false_cnt : 898
+TOTAL:                             5091        63473        5667
+Detected Credentials: 5870
+credsweeper result_cnt : 5165, lost_cnt : 0, true_cnt : 4264, false_cnt : 901
 Category                      TP    FP        TN    FN        FPR        FNR       ACC       PRC       RCL        F1
 --------------------------  ----  ----  --------  ----  ---------  ---------  --------  --------  --------  --------
 Authentication Key & Token    60     4        28    17  0.125      0.220779   0.807339  0.9375    0.779221  0.851064
 Generic Secret               973     3       215    83  0.0137615  0.0785985  0.932496  0.996926  0.921402  0.957677
 Generic Token                289     7       596    44  0.0116086  0.132132   0.945513  0.976351  0.867868  0.918919
-Other                        589   749     63180   255  0.0117161  0.302133   0.9845    0.440209  0.697867  0.539872
+Other                        589   750     63179   255  0.0117318  0.302133   0.984484  0.439881  0.697867  0.539624
 Password                     996   129      4145   406  0.0301825  0.289586   0.905743  0.885333  0.710414  0.788287
 Predefined Pattern           312     2        40    18  0.0476191  0.0545455  0.946237  0.993631  0.945455  0.968944
-Private Key                 1008     2        32     0  0.0588235             0.998081  0.99802   1         0.999009
+Private Key                 1009     4        30     0  0.117647              0.996165  0.996051  1         0.998022
 Seed, Salt, Nonce             36     2         6     4  0.25       0.1        0.875     0.947368  0.9       0.923077
-                            4263   898  19413633   827  4.625e-05  0.162475   0.999911  0.826003  0.837525  0.831724
+                            4264   901  19413629   827  4.641e-05  0.162444   0.999911  0.825557  0.837556  0.831513
